+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Add filtering client function
 - Add sync client function
+- Add createRoom client function
 - Add retryWithLog and syncPoll utility function
 - Add mkReply helper utility function
 - Add reply and edit Event